# specs/001-pilot-space-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pilot Space API
  description: |
    AI-augmented SDLC platform API for project management with intelligent features.

    ## Authentication
    All endpoints require Bearer token authentication via Supabase Auth (GoTrue).
    Authorization is enforced via Row-Level Security (RLS) policies.

    ## Rate Limits
    - Standard endpoints: 1000 requests/minute per workspace
    - AI endpoints: 100 requests/minute per workspace (subject to LLM provider limits)
    - Webhook endpoints: 500 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination using `cursor` and `limit` parameters.
    Response format: `{ data: [...], meta: { total, cursor, hasMore } }`

    ## Updated
    Session 2026-01-22: Migrated to Supabase platform
  version: 2.0.0
  contact:
    name: Pilot Space Team
    url: https://pilot-space.com
  license:
    name: MIT
servers:
  - url: https://api.pilot-space.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
...
paths:
...
  /workspaces/{workspaceSlug}/projects/{projectId}/cycles:
    parameters:
      - $ref: "#/components/parameters/workspaceSlug"
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [Cycles]
      summary: List cycles
      operationId: listCycles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cycle"

    post:
      tags: [Cycles]
      summary: Create cycle
      operationId: createCycle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleRequest"
      responses:
        "201":
          description: Created cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"

  # ============================================================================
  # Modules
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceSlug:
      name: workspaceSlug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Cycle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [upcoming, current, completed]
        issueCount:
          type: integer
        completedIssueCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateCycleRequest:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    # ==================== Modules ====================
...